fix: move ShapeFromVtp to transforms/shape.ts, matching PLY/OBJ pattern
ShapeFromPly and ShapeFromObj live in transforms/shape.ts (moved there by
the upstream OBJ format commit). ShapeFromVtp was wrongly placed in model.ts
during the rebase conflict resolution. Move it to the correct file and update
the format provider to reference StateTransforms.Shape.ShapeFromVtp.
