+++ src/mol-plugin-state/transforms/model.ts
@@ -10,9 +10,7 @@
 import { parseDcd } from '../../mol-io/reader/dcd/parser';
 import { parseGRO } from '../../mol-io/reader/gro/parser';
 import { parsePDB } from '../../mol-io/reader/pdb/parser';
-import { Mat4, Vec3 } from '../../mol-math/linear-algebra';
-import { shapeFromPly } from '../../mol-model-formats/shape/ply';
-import { shapeFromVtp } from '../../mol-model-formats/shape/vtp';
+import { Vec3 } from '../../mol-math/linear-algebra';
 import { coordinatesFromDcd } from '../../mol-model-formats/structure/dcd';
 import { trajectoryFromGRO } from '../../mol-model-formats/structure/gro';
 import { trajectoryFromCCD, trajectoryFromMmCIF } from '../../mol-model-formats/structure/mmcif';
@@ -94,8 +92,6 @@ export { StructureComplexElement };
 export { StructureComponent };
 export { CustomModelProperties };
 export { CustomStructureProperties };
-export { ShapeFromPly };
-export { ShapeFromVtp };
 
 type CoordinatesFromDcd = typeof CoordinatesFromDcd
 const CoordinatesFromDcd = PluginStateTransform.BuiltIn({
@@ -1304,48 +1300,3 @@ async function attachStructureProps(structure: Structure, ctx: PluginContext, ta
     }
 }
 
-type ShapeFromPly = typeof ShapeFromPly
-const ShapeFromPly = PluginStateTransform.BuiltIn({
-    name: 'shape-from-ply',
-    display: { name: 'Shape from PLY', description: 'Create Shape from PLY data' },
-    from: SO.Format.Ply,
-    to: SO.Shape.Provider,
-    params(a) {
-        return {
-            transforms: PD.Optional(PD.Value([Mat4.identity()], { isHidden: true })),
-            label: PD.Optional(PD.Text('', { isHidden: true }))
-        };
-    }
-})({
-    apply({ a, params }) {
-        return Task.create('Create shape from PLY', async ctx => {
-            const shape = await shapeFromPly(a.data, params).runInContext(ctx);
-            const props = { label: params.label || 'Shape' };
-            return new SO.Shape.Provider(shape, props);
-        });
-    }
-});
-
-const _vtpIdentityTransforms = [Mat4.identity()];
-
-type ShapeFromVtp = typeof ShapeFromVtp
-const ShapeFromVtp = PluginStateTransform.BuiltIn({
-    name: 'shape-from-vtp',
-    display: { name: 'Shape from VTP', description: 'Create Shape from VTP (VTK PolyData) file' },
-    from: SO.Format.Vtp,
-    to: SO.Shape.Provider,
-    params(a) {
-        return {
-            transforms: PD.Optional(PD.Value(_vtpIdentityTransforms, { isHidden: true })),
-            label: PD.Optional(PD.Text('', { isHidden: true }))
-        };
-    }
-})({
-    apply({ a, params }) {
-        return Task.create('Create shape from VTP', async ctx => {
-            const shape = await shapeFromVtp(a.data, params).runInContext(ctx);
-            const props = { label: params.label || 'VTP Shape' };
-            return new SO.Shape.Provider(shape, props);
-        });
-    }
-});
